+++ similarity/src/main/java/com/outofprintmagazine/nlp/scorers/descriptive/Questions.java
@@ -20,16 +20,16 @@ public class Questions extends ScorerImpl implements SentenceDescriptiveScorer {
 	}
 	
 	@Override
-	public List<Integer> scoreSentence(CoreSentence sentence) {
-		ArrayList<Integer> retval = new ArrayList<Integer>();
+	public List<Double> scoreSentence(CoreSentence sentence) {
+		ArrayList<Double> retval = new ArrayList<Double>();
 		List<CoreLabel> tokens = sentence.tokens();
 		boolean shouldAdd = tokens.get(tokens.size()-1).lemma().equals("?");
 		for (int i = 0; i < tokens.size(); i++) {
 			if (shouldAdd) {
-				retval.add(new Integer(1));
+				retval.add(new Double(1));
 			}
 			else {
-				retval.add(new Integer(0));
+				retval.add(new Double(0));
 			}				
 		}
 		return retval;
